+++ dart/integration_test/samples/mixins/lib/empty.dart
@@ -1,0 +1,3 @@
+mixin BasicMixin {}
+
+base mixin BaseMixin {}