+++ dart/integration_test/test_utils.dart
@@ -1,8 +1,10 @@
 import 'dart:io';
 
 import 'package:jqassistant_dart_lce/src/core/concepts/class_concept.dart';
+import 'package:jqassistant_dart_lce/src/core/concepts/enum_concept.dart';
 import 'package:jqassistant_dart_lce/src/core/concepts/function_concept.dart';
 import 'package:jqassistant_dart_lce/src/core/concepts/library_concept.dart';
+import 'package:jqassistant_dart_lce/src/core/concepts/mixin_concept.dart';
 import 'package:jqassistant_dart_lce/src/core/concepts/parameter_concept.dart';
 import 'package:jqassistant_dart_lce/src/core/concepts/variable_concept.dart';
 import 'package:path/path.dart';
@@ -77,3 +79,21 @@ Matcher matchesVariable(
         LCEVariable(fqn, libraryPath, name, type, lateModifier, finalModifier,
                 constModifier)
             .toString());
+
+Matcher matchesMixin(
+  String fqn,
+  String libraryPath,
+  String name, {
+  bool baseModifier = false,
+}) =>
+    predicate((LCEMixin v) {
+      return v.fqn == fqn &&
+          v.libraryPath == libraryPath &&
+          v.name == name &&
+          v.baseModifier == baseModifier;
+    }, LCEMixin(fqn, libraryPath, name, baseModifier).toString());
+
+Matcher matchesEnum(String fqn, String libraryPath, String name) =>
+    predicate((LCEEnum v) {
+      return v.fqn == fqn && v.libraryPath == libraryPath && v.name == name;
+    }, LCEEnum(fqn, libraryPath, name).toString());