+++ dart/lib/src/core/utils/ast_utils.dart
@@ -6,8 +6,10 @@
 enum AstNodeType {
   classDeclaration,
   compilationUnit,
+  enumDeclaration,
   formalParameterList,
   functionDeclaration,
   functionExpression,
+  mixinDeclaration,
   topLevelVariableDeclaration,
 }