+++ java/src/main/java/org/jqassistant/plugin/dart/api/model/core/EnumDescriptor.java
@@ -1,0 +1,16 @@
+package org.jqassistant.plugin.dart.api.model.core;
+
+import com.buschmais.xo.neo4j.api.annotation.Label;
+
+@Label("Enum")
+public interface EnumDescriptor extends DartDescriptor, FqnDescriptor {
+
+    String getLibraryPath();
+
+    void setLibraryPath(String libraryPath);
+
+    String getName();
+
+    void setName(String name);
+
+}