+++ java/src/main/java/org/jqassistant/plugin/dart/api/model/core/LibraryDescriptor.java
@@ -17,4 +17,10 @@ public interface LibraryDescriptor extends DartDescriptor, FqnDescriptor {
     @Relation("DECLARES")
     List<VariableDescriptor> getVariables();
 
+    @Relation("DECLARES")
+    List<MixinDescriptor> getMixins();
+
+    @Relation("DECLARES")
+    List<EnumDescriptor> getEnums();
+
 }