+++ java/src/main/java/org/jqassistant/plugin/dart/impl/mapper/core/LibraryMapper.java
@@ -4,10 +4,7 @@ import com.buschmais.jqassistant.core.scanner.api.Scanner;
 import com.buschmais.jqassistant.plugin.common.api.model.FileDescriptor;
 import com.buschmais.jqassistant.plugin.common.api.scanner.FileResolver;
 import lombok.extern.slf4j.Slf4j;
-import org.jqassistant.plugin.dart.api.model.core.ClassDescriptor;
-import org.jqassistant.plugin.dart.api.model.core.FunctionDescriptor;
-import org.jqassistant.plugin.dart.api.model.core.LibraryDescriptor;
-import org.jqassistant.plugin.dart.api.model.core.VariableDescriptor;
+import org.jqassistant.plugin.dart.api.model.core.*;
 import org.jqassistant.plugin.dart.impl.model.ConceptCollection;
 import org.jqassistant.plugin.dart.impl.model.core.Library;
 
@@ -46,6 +43,20 @@ public class LibraryMapper {
                 return o;
             }));
 
+        Map<String, List<MixinDescriptor>> mixins = new HashMap<>();
+        MixinMapper.INSTANCE.mapList(conceptCollection.getMixins(), scanner)
+            .forEach(mixinDescriptor -> mixins.merge(mixinDescriptor.getLibraryPath(), new ArrayList<>(List.of(mixinDescriptor)), (o, n) -> {
+                o.addAll(n);
+                return o;
+            }));
+
+        Map<String, List<EnumDescriptor>> enums = new HashMap<>();
+        EnumMapper.INSTANCE.mapList(conceptCollection.getEnums(), scanner)
+            .forEach(enumDescriptor -> enums.merge(enumDescriptor.getLibraryPath(), new ArrayList<>(List.of(enumDescriptor)), (o, n) -> {
+                o.addAll(n);
+                return o;
+            }));
+
         for (Library library : conceptCollection.getLibraries()) {
             FileDescriptor fileDescriptor = fileResolver.require(library.getPath(), FileDescriptor.class, scanner.getContext());
             if (fileDescriptor != null) { // only represent libraries in the graph that were previously scanned in the file system
@@ -55,6 +66,8 @@ public class LibraryMapper {
                 libraryDescriptor.getClasses().addAll(classes.getOrDefault(library.getPath(), new ArrayList<>()));
                 libraryDescriptor.getFunctions().addAll(functions.getOrDefault(library.getPath(), new ArrayList<>()));
                 libraryDescriptor.getVariables().addAll(variables.getOrDefault(library.getPath(), new ArrayList<>()));
+                libraryDescriptor.getMixins().addAll(mixins.getOrDefault(library.getPath(), new ArrayList<>()));
+                libraryDescriptor.getEnums().addAll(enums.getOrDefault(library.getPath(), new ArrayList<>()));
 
                 result.add(libraryDescriptor);
             }