+++ java/src/test/java/org/jqassistant/plugin/dart/core/basics/DartScannerCoreBasicsIT.java
@@ -62,7 +62,9 @@ class DartScannerCoreBasicsIT extends AbstractPluginIT {
             .assertLibraryPresence()
             .assertClassPresence()
             .assertFunctionPresence()
-            .assertVariablePresence();
+            .assertVariablePresence()
+            .assertMixinPresence()
+            .assertEnumPresence();
 
         store.commitTransaction();
     }