+++ java/src/test/resources/java-it-core-basics-sample-package/lib/enums.dart
@@ -1,0 +1,5 @@
+enum BasicEnum {
+  A,
+  B,
+  C;
+}